+++ src/SharpLearning.Benchmarks/RegressionLearnerBenchmark.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using BenchmarkDotNet.Attributes;
 using SharpLearning.Common.Interfaces;
 using SharpLearning.Containers.Matrices;
@@ -23,8 +22,8 @@ public abstract class RegressionLearnerBenchmark(
     public void GlobalSetup()
     {
         var seed = 42;
-        m_targets = GenerateData(Rows, 1, seed);
-        var features = GenerateData(Rows, Cols, seed);
+        m_targets = DataGenerator.GenerateDoubles(Rows, 1, seed);
+        var features = DataGenerator.GenerateDoubles(Rows, Cols, seed);
         m_features = new F64Matrix(features, Rows, Cols);
 
         // Use default parameters.
@@ -35,12 +34,5 @@ public abstract class RegressionLearnerBenchmark(
     public void Learn()
     {
         m_learner.Learn(m_features, m_targets);
-    }
-
-    public static double[] GenerateData(int rows, int cols, int seed)
-    {
-        var random = new Random(seed);
-        return Enumerable.Range(0, rows * cols)
-            .Select(i => random.NextDouble()).ToArray();
     }
 }
